chore: extract views directory
