+++ apps/client/src/views/timeline/Timeline.module.scss
@@ -1,4 +1,4 @@
-@use '../../../theme/viewerDefs' as *;
+@use '../../theme/viewerDefs' as *;
 
 $timeline-height: 1rem;
 $view-background: $ui-black;
