+++ apps/client/src/views/timeline/TimelinePage.scss
@@ -1,4 +1,4 @@
-@use '../../../theme/viewerDefs' as *;
+@use '../../theme/viewerDefs' as *;
 
 .timeline {
   width: 100vw;
