+++ apps/client/src/views/timeline/timeline.options.ts
@@ -1,5 +1,5 @@
-import { getTimeOption } from '../../../common/components/view-params-editor/constants';
-import { ViewOption } from '../../../common/components/view-params-editor/types';
+import { getTimeOption } from '../../common/components/view-params-editor/constants';
+import { ViewOption } from '../../common/components/view-params-editor/types';
 
 export const getTimelineOptions = (timeFormat: string): ViewOption[] => {
   return [
